core/requirements.txt
asgiref==3.8.1
astroid==3.3.10
astunparse==1.6.3
attrs==25.3.0
CacheControl==0.14.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
cryptography==44.0.1
Cython==3.0.12
dill==0.4.0
dj-database-url==2.3.0
Django==5.1.6
django-cors-headers==4.7.0
django-environ==0.12.0
django-extensions==4.1
django-filter==25.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.8
firebase-admin==6.6.0
google-api-core==2.24.1
google-api-python-client==2.162.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.2
google-cloud-firestore==2.20.1
google-cloud-storage==3.1.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.68.0
graphviz==0.20.3
grpcio==1.70.0
grpcio-status==1.70.0
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
inflection==0.5.1
isort==6.0.1
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lxml==5.4.0
MarkupSafe==3.0.2
mccabe==0.7.0
msgpack==1.1.0
packaging==24.2
pillow==11.1.0
platformdirs==4.3.8
proto-plus==1.26.0
protobuf==5.29.3
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydotplus==2.0.2
PyJWT==2.9.0
pylint==3.3.7
pyparsing==3.2.1
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
rsa==4.9
setuptools==75.8.0
six==1.17.0
sqlparse==0.5.3
stripe==11.6.0
tomlkit==0.13.2
typing_extensions==4.13.0
uritemplate==4.1.1
urllib3==2.3.0
wheel==0.45.1
whitenoise==6.9.0
